--- src/components/ListVideosCard/index.jsx
@@ -1,1 +1,0 @@
-export { default } from './ListVideosCard.component';
