+++ src/components/Videos/VideoCard/index.jsx
@@ -1,0 +1,1 @@
+export { default } from './VideoCard.component';
